requirements.txt
pyyaml
requests
xdg
